backend/requirements.txt
# OracleDota Backend Dependencies

# Web Framework
fastapi
uvicorn[standard]
gunicorn

# HTTP Requests
requests
# axios removed (frontend lib)

# Environment Variables
python-dotenv

# AI & LLM (DeepSeek via OpenRouter)
openai

# Database & Cloud
firebase-admin

# TTS (Voice)
edge-tts
aiofiles

# RAG 2.0 (Vector DB)
faiss-cpu
sentence-transformers
numpy
